AoC 2022: day 17 - mathematically not 100% correct version used to submit results
I'm unable to prove why the wrap-detection like this is always correct,
it seems to me this could actually fail for precisely crafted input.

To fix the tracked state for wrap detection should include also chamber
state. Most of the solutions get away with heuristic tracking only few
top lines of chamber, but even that is more than this solution.

This solution is based only on assumption that when piece == 0 with same
input ahead (ignoring first input.length * 2 pieces) is processed, the
tower added will be identical to previous. As the condition is verified
for all possible jet (input index value) checking they all grow at the
same speed. This is probably near impossible to break (with non-trivial
case), but I can't prove it, so TODO is to include some chamber state
too.