AoC 2022: day 19 - fix wrong init and exit of simulation
the simulation was starting with already ore=1, kinda fast-forwarding
first minute, and compensating for it by stopping simulation one turn
earlier.

But this was not intentional, and done in a way affecting the early-exit
pruning rule in wrong way, causing it to be about 4-5 times slower than
this fixed version.

This is still ultra slow, requires 1min for sample.txt and 13:25 for
input.txt on my machine, because this is very naive per-minute DFS
trying out also lot of meaningless permutations (like trying both
variants of building particular type of robot and waiting one more
minute before building the same type of robot, which gives no advantage
ever, only waiting for different types makes sense).